+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0](https://github.com/sinarionn/ReusableView/releases/tag/2.0.0)
+
+Naming changes.
+DistinctiveReuse - removed.
+rx.reuseBag - setter removed.
+rx.viewModel - binder added.
+
+
 ## [1.2.1](https://github.com/sinarionn/ReusableView/releases/tag/1.2.1)
 
 Fixed bug with structures hidden behind protocols